crml_engine/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crml-engine"
version = "1.1.0"
description = "CRML reference engine (simulation/runtime)"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  {name = "Zeron Research Labs"},
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "crml-lang>=1.1.0",
  "numpy",
  "scipy",
  "pydantic",
  "pyyaml",
]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
crml = "crml_engine.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
